Remove no need public function from StaticPress
